+++ NEWS.md
@@ -1,6 +1,7 @@
 # parquetize 0.2.0
 
-* Added `json_to_parquet()` function to convert json and ndjson files to parquet format
+* Added `json_to_parquet()` function to convert json and ndjson files to parquet format.
+* Added the possibility to convert a csv file to a partitioned parquet file.
 * Improving code coverage (#9)
 * Check if `path_to_parquet` exists in functions `csv_to_parquet()` or `table_to_parquet()` (@py-b)